[issue-6] extract "low level" PNG functions into LowLevel module
