+++ apps/png/lib/png.ex
@@ -3,16 +3,7 @@ defmodule PNG do
   Documentation for PNG.
   """
 
-  alias PNG.Config
-
-  @type chunk ::
-          {:raw, iodata()}
-          | {:row, iodata()}
-          | {:rows, iodata()}
-          | {:data, iodata()}
-          | {:compressed, iodata()}
-
-  @scanline_filter 0
+  alias PNG.{Config, LowLevel}
 
   @spec create(map()) :: map()
   def create(%{file: file} = png) do
@@ -23,8 +14,8 @@ defmodule PNG do
 
   def create(%{size: {width, height} = size, mode: mode, call: callback} = png) do
     config = %Config{size: {width, height}, mode: mode}
-    :ok = callback.(header())
-    :ok = callback.(chunk("IHDR", config))
+    :ok = callback.(LowLevel.header())
+    :ok = callback.(LowLevel.chunk("IHDR", config))
     :ok = append_palette(png)
     z = :zlib.open()
     :ok = :zlib.deflateInit(z)
@@ -32,7 +23,7 @@ defmodule PNG do
   end
 
   def append_palette(%{call: callback, palette: palette}) do
-    chunk = chunk("PLTE", palette)
+    chunk = LowLevel.chunk("PLTE", palette)
     :ok = callback.(chunk)
   end
 
@@ -40,7 +31,7 @@ defmodule PNG do
     :ok
   end
 
-  @spec append(map(), chunk()) :: map()
+  @spec append(map(), LowLevel.chunk()) :: map()
   def append(png, {:row, row}) do
     append(png, {:data, [0, row]})
   end
@@ -61,7 +52,7 @@ defmodule PNG do
   end
 
   def append(%{call: callback} = png, {:compressed, compressed}) do
-    chunks = chunk("IDAT", {:compressed, compressed})
+    chunks = LowLevel.chunk("IDAT", {:compressed, compressed})
     :ok = callback.(chunks)
     png
   end
@@ -72,100 +63,7 @@ defmodule PNG do
     append(png, {:compressed, List.flatten(compressed)})
     :ok = :zlib.deflateEnd(z)
     :ok = :zlib.close(z)
-    :ok = callback.(chunk("IEND"))
+    :ok = callback.(LowLevel.chunk("IEND"))
     :ok
-  end
-
-  @doc """
-  Returns the header for a PNG file.
-  """
-  @spec header :: binary()
-  def header do
-    <<0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A>>
-  end
-
-  @doc """
-  blah blah blah
-
-  This supports only basic compression, filter, and interlace methods.
-  Only supports scanline filter 0 (i.e., none).
-  """
-  @spec chunk(String.t(), Config.t() | chunk() | binary()) :: binary()
-  def chunk(type, data \\ <<>>)
-
-  def chunk(
-        "IHDR",
-        %Config{
-          size: {width, height},
-          mode: {color_type, bit_depth},
-          compression_method: 0,
-          filter_method: 0,
-          interlace_method: 0
-        } = config
-      ) do
-    color_byte_type =
-      case color_type do
-        :grayscale -> 0
-        :rgb -> 2
-        :indexed -> 3
-        :grayscale_alpha -> 4
-        :rgba -> 6
-      end
-
-    data = <<
-      width::32,
-      height::32,
-      bit_depth::8,
-      color_byte_type::8,
-      config.compression_method::8,
-      config.filter_method::8,
-      config.interlace_method::8
-    >>
-
-    chunk(<<"IHDR">>, data)
-  end
-
-  def chunk("IDAT", {:rows, rows}) do
-    raw = :erlang.list_to_binary(for row <- rows, do: [@scanline_filter, row])
-    chunk("IDAT", {:raw, raw})
-  end
-
-  def chunk("IDAT", {:raw, data}) do
-    chunk("IDAT", {:compressed, compress(data)})
-  end
-
-  def chunk("IDAT", {:compressed, compressed_data}) when is_list(compressed_data) do
-    Enum.map(compressed_data, fn part ->
-      chunk(<<"IDAT">>, part)
-    end)
-  end
-
-  def chunk("PLTE", {:rgb, bit_depth, color_tuples}) do
-    data =
-      color_tuples
-      |> Enum.map(fn {r, g, b} ->
-        <<r::size(bit_depth), g::size(bit_depth), b::size(bit_depth)>>
-      end)
-      |> :erlang.list_to_binary()
-
-    chunk("PLTE", data)
-  end
-
-  def chunk(type, data) when is_binary(type) and is_binary(data) do
-    length = byte_size(data)
-    type_data = <<type::binary, data::binary>>
-    crc = :erlang.crc32(type_data)
-    <<length::32, type_data::binary, crc::32>>
-  end
-
-  @spec compress(binary()) :: [binary()]
-  def compress(data) do
-    with z = :zlib.open(),
-         :ok <- :zlib.deflateInit(z),
-         compressed = :zlib.deflate(z, data, :finish),
-         :ok <- :zlib.deflateEnd(z),
-         :ok <- :zlib.close(z) do
-      List.flatten(compressed)
-    end
   end
 end
